now its cub3d

+++ include/utils.h
@@ -6,7 +6,7 @@
 /*   By: wdavey <wdavey@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/29 13:54:30 by wdavey            #+#    #+#             */
-/*   Updated: 2023/08/31 15:57:31 by wdavey           ###   ########.fr       */
+/*   Updated: 2024/02/01 13:54:42 by wdavey           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,10 @@
 # include <stdbool.h>
 # include <stddef.h>
 
-typedef struct s_pos {
-	long	x;
-	long	y;
+typedef struct s_pos
+{
+	float	x;
+	float	y;
 }	t_pos;
 
 t_pos	pos_new(long x, long y);
